requirements.txt
importlib-metadata>=4.8.2
importlib-resources>=5.4.0
ipykernel>=6.5.0
ipython-genutils>=0.2.0
ipython>=7.29.0
ipywidgets>=7.6.5
joblib>=1.1.0
matplotlib-inline>=0.1.3
matplotlib>=3.5.0
notebook>=6.4.5
numpy>=1.19.5
pytorch-lightning>=1.6.0
scikit-learn-extra>=0.2.0
scikit-learn>=1.0.1
seaborn>=0.11.2
torch>=1.11.0
torchmetrics>=0.6.2
torchvision>=0.12.0